Refactor main.ts
