+++ src/common/utils/history.ts
@@ -1,0 +1,7 @@
+import { createMemoryHistory } from 'history'
+import * as ROUTES from 'common/utils/constants/routes'
+
+export default createMemoryHistory({
+  initialEntries: [ROUTES.LOADING],
+  initialIndex: 0,
+})
