+++ src/features/deepLinking/index.ts
@@ -1,13 +1,1 @@
-import {
-  forceSingleInstanceApplication,
-  redirectDeepLinkingUrl,
-  registerCustomProtocol,
-  registerFileProtocol,
-} from './DeepLinking.utils'
-
-export {
-  forceSingleInstanceApplication,
-  redirectDeepLinkingUrl,
-  registerCustomProtocol,
-  registerFileProtocol,
-}
+export * from './DeepLinking.utils'
